+++ src/main/java/org/truffleruby/core/mutex/MutexNodes.java
@@ -9,7 +9,6 @@
  */
 package org.truffleruby.core.mutex;
 
-import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
 import com.oracle.truffle.api.dsl.Cached;
 import com.oracle.truffle.api.dsl.CreateCast;
 import com.oracle.truffle.api.dsl.NodeChild;
@@ -102,20 +101,9 @@ public abstract class MutexNodes {
             if (heldByCurrentThreadProfile.profile(lock.isHeldByCurrentThread())) {
                 return false;
             } else {
-                return doTryLock(thread, lock);
+                return MutexOperations.tryLock(lock, thread);
             }
         }
-
-        @TruffleBoundary
-        private boolean doTryLock(RubyThread thread, ReentrantLock lock) {
-            if (lock.tryLock()) {
-                thread.ownedLocks.add(lock);
-                return true;
-            } else {
-                return false;
-            }
-        }
-
     }
 
     @CoreMethod(names = "unlock")
@@ -150,6 +138,9 @@ public abstract class MutexNodes {
                 throw new RaiseException(getContext(), coreExceptions().threadErrorRecursiveLocking(this));
             }
 
+            /* This code uses lock/unlock because the list of owned locks must be maintained. User code can unlock a
+             * mutex inside a synchronize block, and then relock it before exiting the block, and we need the owned
+             * locks list to be in consistent state at the end. */
             MutexOperations.lock(getContext(), lock, thread, this);
             try {
                 return callBlock(block);